+++ tests/unit/test_application.py
@@ -23,7 +23,6 @@ import pathlib
 import re
 import subprocess
 import sys
-from datetime import date
 from textwrap import dedent
 from unittest import mock
 
@@ -592,79 +591,6 @@ def test_run_success_unmanaged(
         emitter.assert_debug("Preparing application...")
     with check:
         emitter.assert_debug("Running testcraft pass on host")
-
-
-@pytest.mark.skipif(
-    date.today() <= date(2025, 3, 31), reason="run_managed is going away."
-)
-def test_run_success_managed(monkeypatch, app, fake_project, mocker):
-    mocker.patch.object(app, "get_project", return_value=fake_project)
-    app.run_managed = mock.Mock()
-    monkeypatch.setattr(sys, "argv", ["testcraft", "pull"])
-
-    pytest_check.equal(app.run(), 0)
-
-    app.run_managed.assert_called_once_with(None, None)  # --build-for not used
-
-
-@pytest.mark.skipif(
-    date.today() <= date(2025, 3, 31), reason="run_managed is going away."
-)
-def test_run_success_managed_with_arch(monkeypatch, app, fake_project, mocker):
-    mocker.patch.object(app, "get_project", return_value=fake_project)
-    app.run_managed = mock.Mock()
-    arch = get_host_architecture()
-    monkeypatch.setattr(sys, "argv", ["testcraft", "pull", f"--build-for={arch}"])
-
-    pytest_check.equal(app.run(), 0)
-
-    app.run_managed.assert_called_once()
-
-
-@pytest.mark.skipif(
-    date.today() <= date(2025, 3, 31), reason="run_managed is going away."
-)
-def test_run_success_managed_with_platform(monkeypatch, app, fake_project, mocker):
-    mocker.patch.object(app, "get_project", return_value=fake_project)
-    app.run_managed = mock.Mock()
-    monkeypatch.setattr(sys, "argv", ["testcraft", "pull", "--platform=foo"])
-
-    pytest_check.equal(app.run(), 0)
-
-    app.run_managed.assert_called_once_with("foo", None)
-
-
-@pytest.mark.skipif(
-    date.today() <= date(2025, 3, 31), reason="run_managed is going away."
-)
-@pytest.mark.parametrize(
-    ("params", "expected_call"),
-    [
-        ([], mock.call(None, None)),
-        (["--platform=s390x"], mock.call("s390x", None)),
-        (
-            ["--platform", get_host_architecture()],
-            mock.call(get_host_architecture(), None),
-        ),
-        (
-            ["--build-for", get_host_architecture()],
-            mock.call(None, get_host_architecture()),
-        ),
-        (["--build-for", "s390x"], mock.call(None, "s390x")),
-        (["--platform", "s390x,riscv64"], mock.call("s390x", None)),
-        (["--build-for", "s390x,riscv64"], mock.call(None, "s390x")),
-    ],
-)
-def test_run_passes_platforms(
-    monkeypatch, app, fake_project, mocker, params, expected_call
-):
-    mocker.patch.object(app, "get_project", return_value=fake_project)
-    app.run_managed = mock.Mock(return_value=False)
-    monkeypatch.setattr(sys, "argv", ["testcraft", "pull", *params])
-
-    pytest_check.equal(app.run(), 0)
-
-    assert app.run_managed.mock_calls == [expected_call]
 
 
 @pytest.mark.parametrize("return_code", [None, 0, 1])